Edits for eventageous

+++ components/events/EventCard.js
@@ -1,8 +1,7 @@
 import PropTypes from 'prop-types';
-import { useContext } from 'react';
+import { useContext, useState, useEffect } from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import { useState, useEffect } from 'react';
 import { format } from 'date-fns';
 import { Card, Button } from 'react-bootstrap';
 import OrderContext from '../../utils/context/orderContext';
@@ -35,15 +34,19 @@ export default function EventCard({ event }) {
     });
   };
 
-  const handleDelete = async () => {
-    deleteEvent(event.id).then(router.push('/myevents'));
+  const handleDelete = () => {
+    if (window.confirm(`Delete ${event.name}?`)) {
+      deleteEvent(event.id).then(router.push('/myevents'));
+    }
   };
 
   return (
     <div className="card">
       <Card className="event-card" style={{ width: '300px' }}>
         <Card.Header>{event.name}</Card.Header>
-        <Card.Img variant="top" className="event-card-img" src={event.image_url} style={{ width: '200px' }} />
+        <div className="card-img-container">
+          <Card.Img variant="top" className="event-card-img" src={event.image_url} style={{ width: '100%' }} />
+        </div>
         <Card.Body>
           <Card.Text>{event.description}</Card.Text>
           <Card.Text>Ticket Price: ${event.ticket?.price}</Card.Text>
@@ -90,9 +93,11 @@ EventCard.propTypes = {
     name: PropTypes.string,
     description: PropTypes.string,
     price: PropTypes.number,
-    availableQuantity: PropTypes.number,
-    imageUrl: PropTypes.string,
+    tickets_available: PropTypes.number,
+    image_url: PropTypes.string,
+    date: PropTypes.string,
     ticket: PropTypes.shape({
+      id: PropTypes.number,
       price: PropTypes.string,
     }),
     seller: PropTypes.shape({
